# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Fleet Manager API",
    "version": "1.0.0",
    "description": "API documentation for the Fleet Manager application"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local server"
    }
  ],
  "paths": {
    "/users/create": {
      "post": {
        "summary": "Create a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "username", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "500": {
            "description": "Error creating user"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "summary": "Login a user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "404": {
            "description": "User not found"
          },
          "400": {
            "description": "Invalid credentials"
          },
          "500": {
            "description": "Error logging in"
          }
        }
      }
    },
    "/drivers": {
      "get": {
        "summary": "Get all drivers",
        "responses": {
          "200": {
            "description": "List of all drivers"
          },
          "500": {
            "description": "Error fetching drivers"
          }
        }
      },
      "post": {
        "summary": "Create a new driver",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "licenseNumber": { "type": "string" },
                  "phone": { "type": "string" }
                },
                "required": ["name", "licenseNumber", "phone"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Driver created successfully"
          },
          "500": {
            "description": "Error creating driver"
          }
        }
      }
    },
    "/drivers/{id}": {
      "get": {
        "summary": "Get a driver by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Driver fetched successfully"
          },
          "404": {
            "description": "Driver not found"
          },
          "500": {
            "description": "Error fetching driver"
          }
        }
      },
      "put": {
        "summary": "Update a driver by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "licenseNumber": { "type": "string" },
                  "phone": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Driver updated successfully"
          },
          "404": {
            "description": "Driver not found"
          },
          "500": {
            "description": "Error updating driver"
          }
        }
      },
      "delete": {
        "summary": "Delete a driver by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Driver deleted successfully"
          },
          "404": {
            "description": "Driver not found"
          },
          "500": {
            "description": "Error deleting driver"
          }
        }
      }
    }
  }
}